+++ src/Event/AbstractEvent.php
@@ -2,8 +2,8 @@
 
 namespace TK\GitHubWebhook\Event;
 
-use TK\GitHubWebhook\Model\Repository;
-use TK\GitHubWebhook\Model\User;
+use TK\GitHubWebhook\Model\Common\Repository;
+use TK\GitHubWebhook\Model\Common\User;
 
 abstract class AbstractEvent
 {