+++ src/Event/StarEvent.php
@@ -2,10 +2,10 @@
 
 namespace TK\GitHubWebhook\Event;
 
-use TK\GitHubWebhook\Model\InstallationLite;
-use TK\GitHubWebhook\Model\Repository;
+use TK\GitHubWebhook\Model\Common\InstallationLite;
+use TK\GitHubWebhook\Model\Common\Repository;
 use TK\GitHubWebhook\Model\Star\EventTypes;
-use TK\GitHubWebhook\Model\User;
+use TK\GitHubWebhook\Model\Common\User;
 use TK\GitHubWebhook\Util;
 
 class StarEvent extends AbstractEvent