+++ src/Model/Issue/Changes.php
@@ -2,8 +2,9 @@
 
 namespace TK\GitHubWebhook\Model\Issue;
 
-use TK\GitHubWebhook\Model\Issue;
-use TK\GitHubWebhook\Model\Repository;
+use TK\GitHubWebhook\Model\Common\From;
+use TK\GitHubWebhook\Model\Common\Issue;
+use TK\GitHubWebhook\Model\Common\Repository;
 use TK\GitHubWebhook\Util;
 
 readonly class Changes